fix: run polyglot SQL parse on large-stack pool
SqlParseCache was calling polyglot-sql on Tokio workers (~2 MiB),
which stack-overflowed on queries like unnest([...]) in snowflake
wire e2e. Reuse the dedicated 16 MiB polyglot pool from core.
